Milestone: Service Clients Sequentially
This commit ensures that multiple clients can be serviced by the runtime. But, this is still done in a sequential manner. This is demonstrated by the presence of the sleep timer in the client state machine which artificially delays the response. All requests are delayed by the sleep time in a sequential manner.